定数に e, π, 2π を追加

+++ src/com/github/novisoftware/patternDraw/gui/editor/core/langSpec/typeSystem/ValueFloat.java
@@ -9,7 +9,7 @@ public class ValueFloat extends ValueAbstractScalar {
 	}
 
 	public ValueFloat(Double d) {
-		super(ValueType.INTEGER);
+		super(ValueType.FLOAT);
 		internal = d;
 	}
